feat: es 查询支持自定义时间字段 #1010158081118404350

+++ pkg/unify-query/tsdb/elasticsearch/format.go
@@ -62,7 +62,7 @@ const (
 
 const (
 	TimeFieldTypeTime = "date"
-	TimeFieldTypeInt  = "int"
+	TimeFieldTypeInt  = "long"
 )
 
 const (
